381: move hero code from actors to hero
